test: Improve ECOnlyTestFixture and use it in sparse_reads_cxx.cc
Use a single pool for all tests in the suite, instead of one pool per test.
Create new ECSparseReadTest class inheriting from this fixture, and use it for tests currently skipping Replicated pools.
Create SparseReadHelpers class giving both sparse read test classes access to the same helper functions.

Fixes: https://tracker.ceph.com/issues/79804
