+++ src/test/librados/test_pool_types.cc
@@ -4,6 +4,7 @@
 #include "test_pool_types.h"
 #include "common/json/OSDStructures.h"
 #include "include/scope_guard.h"
+#include "osd/osd_types.h"
 
 #include <algorithm>
 #include <chrono>
@@ -25,9 +26,13 @@ std::map<PoolType, std::string> PoolTypeTestFixture::pool_names;
 librados::Rados ECOnlyTestFixture::rados;
 std::string ECOnlyTestFixture::static_pool_name;
 
-void PoolTypeTestFixture::cleanup_namespace(librados::Rados& cluster,
-                                            librados::IoCtx& ioctx,
-                                            const std::string& ns) {
+// ---------------------------------------------------------------------------
+// File-scope helper: removes all objects in the given namespace from ioctx.
+// Used by both PoolTypeTestFixture and ECOnlyTestFixture to avoid duplication.
+// ---------------------------------------------------------------------------
+static void do_cleanup_namespace(librados::Rados& cluster,
+                                 librados::IoCtx& ioctx,
+                                 const std::string& ns) {
   ioctx.snap_set_read(librados::SNAP_HEAD);
   ioctx.set_namespace(ns);
 
@@ -55,6 +60,18 @@ void PoolTypeTestFixture::cleanup_namespace(librados::Rados& cluster,
     }
     sleep(1);
   }
+}
+
+void PoolTypeTestFixture::cleanup_namespace(librados::Rados& cluster,
+                                            librados::IoCtx& ioctx,
+                                            const std::string& ns) {
+  do_cleanup_namespace(cluster, ioctx, ns);
+}
+
+void ECOnlyTestFixture::cleanup_namespace(librados::Rados& cluster,
+                                          librados::IoCtx& ioctx,
+                                          const std::string& ns) {
+  do_cleanup_namespace(cluster, ioctx, ns);
 }
 
 void PoolTypeTestFixture::SetUpTestSuite() {
@@ -239,6 +256,19 @@ void PoolTypeTestFixture::setup_and_trigger_recovery(
   // Wait for new upmap to appear as acting set of osds
   int res2 = wait_for_upmap(oid, new_primary, timeout);
   EXPECT_EQ(0, res2);
+}
+
+void ECOnlyTestFixture::SetUpTestSuite() {
+  ASSERT_EQ("", connect_cluster_pp(rados));
+  static_pool_name = get_temp_pool_name("ec_only_test_");
+  ASSERT_EQ("", create_pool_by_type(static_pool_name, rados, PoolType::FAST_EC));
+  rados.wait_for_latest_osdmap();
+}
+
+void ECOnlyTestFixture::TearDownTestSuite() {
+  ASSERT_EQ(0, destroy_pool_by_type(static_pool_name, rados, PoolType::FAST_EC));
+  static_pool_name.clear();
+  rados.shutdown();
 }
 
 void ECOnlyTestFixture::SetUp() {
